requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
aiosqlite==0.20.0
python-dotenv==1.0.1
xmltodict==0.13.0
unidecode==1.3.8
yt-dlp==2024.11.18
musicbrainzngs==0.7.1
playwright==1.49.1
pyyaml==6.0.2
